requirements.txt
#API
fastapi
uvicorn

mysql-connector-python
pydantic-settings